+++ src/DynAI/Interfaces/IInputData.cs
@@ -1,0 +1,20 @@
+﻿using Autodesk.DesignScript.Runtime;
+
+namespace AI
+{
+    [IsVisibleInDynamoLibrary(false)]
+    public interface IInputData<T>
+    {
+        /// <summary>
+        /// Retrieves the input data.
+        /// </summary>
+        /// <returns>The data.</returns>
+        T GetInputData();
+
+        /// <summary>
+        /// Sets the input data for the machine to the specified values.
+        /// </summary>
+        /// <returns>True if the operation succeeded, false otherwise.</returns>
+        bool SetInputData();
+    }
+}
