+++ src/DynAI/Interfaces/IMachine.cs
@@ -4,7 +4,7 @@ using Autodesk.DesignScript.Runtime;
 namespace AI
 {
     /// <summary>
-    /// The generic capabilities of a machine, used for machine learning
+    /// The generic capabilities of a machine that can learn and predict.
     /// </summary>
     [IsVisibleInDynamoLibrary(false)]
     public interface IMachine
@@ -57,40 +57,5 @@ namespace AI
         dynamic Predict(dynamic testData);
 
         #endregion
-    }
-
-    /// <summary>
-    /// This provides an abstraction so that different machines can use different formats for training data.
-    /// </summary>
-    [IsVisibleInDynamoLibrary(false)]
-    public interface ITrainingData
-    {
-        /// <summary>
-        /// Fetches the training data.
-        /// </summary>
-        /// <returns>The training dataset.</returns>
-        double[] GetTrainingData();
-
-        /// <summary>
-        /// Allows us to set the training data.
-        /// </summary>
-        /// <returns>True if the operation succeeded, false otherwise.</returns>
-        bool SetTrainingData();
-    }
-
-    [IsVisibleInDynamoLibrary(false)]
-    public interface IInputData<T>
-    {
-        /// <summary>
-        /// Retrieves the input data.
-        /// </summary>
-        /// <returns>The data.</returns>
-        T GetInputData();
-
-        /// <summary>
-        /// Sets the input data for the machine to the specified values.
-        /// </summary>
-        /// <returns>True if the operation succeeded, false otherwise.</returns>
-        bool SetInputData();
     }
 }
