+++ src/DynAI/Interfaces/ITrainingData.cs
@@ -1,0 +1,28 @@
+﻿using Autodesk.DesignScript.Runtime;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AI
+{
+    /// <summary>
+    /// This provides an abstraction so that different machines can use different formats for training data.
+    /// </summary>
+    [IsVisibleInDynamoLibrary(false)]
+    public interface ITrainingData
+    {
+        /// <summary>
+        /// Fetches the training data.
+        /// </summary>
+        /// <returns>The training dataset.</returns>
+        double[] GetTrainingData();
+
+        /// <summary>
+        /// Allows us to set the training data.
+        /// </summary>
+        /// <returns>True if the operation succeeded, false otherwise.</returns>
+        bool SetTrainingData();
+    }
+}
